+++ addon/components/fd-uml-diagram.js
@@ -210,6 +210,7 @@ export default Component.extend({
     paper.on('updaterepobj', this._updateRepObj, this);
     paper.on('checkexistelements', this._checkOnExistElements, this);
     paper.on('cell:highlight', this._highlighted, this);
+    paper.on('element:openeditform', this._elementOpenEditForm, this);
 
     // Ghost element mode.
     paper.on('element:pointermove', this._ghostElementMove);
@@ -282,6 +283,19 @@ export default Component.extend({
         this._clearLinksData();
       }
     }
+  },
+
+  /**
+    Handler custom event `element:openeditform`.
+    Invokes the action specified in the `openEditFormAction` property, passing it the UML-element model.
+
+    @private
+    @method _elementOpenEditForm
+    @param {joint.dia.CellView} cellView
+    @param {joint.dia.Element} cellView.model
+  */
+  _elementOpenEditForm({ model }) {
+    this.get('openEditFormAction')(model.get('objectModel'));
   },
 
   /**